+++ modules/common/optional/proxy.nix
@@ -5,10 +5,22 @@
   ...
 }:
 let
+  inherit (lib)
+    mkIf
+    ;
   cfg = config.htn3.optional.proxy;
 in
-lib.mkIf (config.htn3.enable && cfg.enable) {
+mkIf (config.htn3.enable && cfg.enable) {
   environment.systemPackages = with pkgs; [ clashtui ];
+
+  services = mkIf cfg.mihomo.enable {
+    mihomo = {
+      enable = true;
+      tunMode = true;
+      inherit (cfg.mihomo) configFile;
+      webui = lib.mkDefault pkgs.metacubexd;
+    };
+  };
 
   # Configure network proxy if necessary
   # networking.proxy.default = "http://user:password@proxy:port/";